setup small package
