+++ .gitignore
@@ -28,3 +28,4 @@ opf.py
 mem_test
 
 GUROBI_RUN*
+nomopyomo.egg-info/PKG-INFO
