+++ nomopyomo/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Sat Sep  7 17:51:14 2019
+
+@author: fabian
+"""
+
+from . import nomopyomo
