+++ setup.py
@@ -1,0 +1,12 @@
+from __future__ import absolute_import
+
+
+
+from setuptools import setup, find_packages
+from codecs import open
+
+
+setup(
+    name='nomopyomo',
+    version='0.1alpha',
+    license='GPLv3')
